requirements.txt
requests
python-dotenv
matplotlib
Pillow
pandas
python-telegram-bot==13.15

